docs 오류 및 자료형 수정
subject : 오류 및 자료형 수정

+++ So1omon/src/main/java/com/kh/so1omon/product/model/dao/ProductDao.java
@@ -124,7 +124,7 @@ public class ProductDao {
 		return (ArrayList)sqlSession.selectList("productMapper.selectMyPageWishList", mno, rowBounds);
 	}
 
-	public ArrayList<Review> selectReviewListAD(SqlSessionTemplate sqlSession, int userNo){
+	public ArrayList<Review> selectReviewListAD(SqlSessionTemplate sqlSession, long userNo){
 		return (ArrayList)sqlSession.selectList("productMapper.selectReviewListAD", userNo);
 	}
